Split code from main.c
