+++ src/clock_configuration.c
@@ -1,5 +1,5 @@
 #include "stm32f4xx.h"
-#include "main.h"
+#include "commands_manager.h"
 
 TIM_HandleTypeDef htim14;
